+++ tests/testthat/test-export-modified.R
@@ -17,3 +17,19 @@ test_that("exporting unnamed lists works", {
   expect_identical(x$y$z, as.list(1:10))
 
 })
+
+
+test_that("exporting undefs works", {
+
+  fn <- function(n) {
+    x <- tf$constant(0L)
+    autograph(for (i in tf$range(tf$cast(n, tf$int32))) {
+      x <- x + i
+    })
+    x
+  }
+
+  tf_fn <- .tf_function(fn)
+  expect_equal(as.vector(tf_fn(tf$constant(5L))), 10L)
+
+})